+++ apps/todo-app-with-auth-backend/queues/hello-world.ts
@@ -1,4 +1,4 @@
-import { ZembleQueue } from 'zemble-plugin-bull'
+import { ZembleQueue } from '@zemble/bull'
 
 export default new ZembleQueue((job, { logger }) => {
   logger.info(job.data)