+++ apps/todo-app-with-auth/App.tsx
@@ -1,9 +1,9 @@
-import LoginButton from 'zemble-plugin-auth-anonymous-expo/components/LoginButton'
-import { ShowForAuthenticated } from 'zemble-plugin-auth-anonymous-expo/components/ShowForAuthenticated'
-import { ShowForUnauthenticated } from 'zemble-plugin-auth-anonymous-expo/components/ShowForUnauthenticated'
-import PluginProvider from 'zemble-plugin-auth-anonymous-expo/contexts/Plugin'
-import { LogoutButton } from 'zemble-plugin-auth-expo/components/LogoutButton'
-import TodoList from 'zemble-plugin-todo-ui/components/TodoList'
+import LoginButton from '@zemble/auth-anonymous-expo/components/LoginButton'
+import { ShowForAuthenticated } from '@zemble/auth-anonymous-expo/components/ShowForAuthenticated'
+import { ShowForUnauthenticated } from '@zemble/auth-anonymous-expo/components/ShowForUnauthenticated'
+import PluginProvider from '@zemble/auth-anonymous-expo/contexts/Plugin'
+import { LogoutButton } from '@zemble/auth-expo/components/LogoutButton'
+import TodoList from '@zemble/todo-ui/components/TodoList'
 
 export default function App() {
   return (