+++ packages/debug/plugin.ts
@@ -1,7 +1,7 @@
+import Auth from '@zemble/auth'
 import { Plugin } from '@zemble/core'
 import GraphQL from '@zemble/graphql'
 import Routes from '@zemble/routes'
-import Auth from 'zemble-plugin-auth'
 
 export default new Plugin(
   import.meta.dir,