+++ packages/debug/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - 67e139d: Add @zemble/debug package
 - Updated dependencies [67e139d]
-  - zemble-plugin-auth@0.0.41
+  - @zemble/auth@0.0.41
   - @zemble/core@0.1.11
   - @zemble/graphql@0.0.42
   - @zemble/bun@0.0.22